@@ -1,10 +1,11 @@
+﻿using System;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Blazored.LocalStorage.JsonConverters;
+using Blazored.LocalStorage.Serialization;
 using Blazored.LocalStorage.StorageOptions;
-using Blazored.LocalStorage.Tests.Mocks;
+using Blazored.LocalStorage.Testing;
 using Blazored.LocalStorage.Tests.TestAssets;
-using FluentAssertions;
 using Microsoft.Extensions.Options;
 using Moq;
 using Xunit;
@@ -13,96 +14,93 @@ namespace Blazored.LocalStorage.Tests.LocalStorageServiceTests
 {
     public class GetItemAsync
     {
-        private readonly JsonSerializerOptions _jsonOptions;
-        private readonly Mock<JSRuntimeWrapperAsync> _mockJSRuntime;
-        private readonly Mock<IOptions<LocalStorageOptions>> _mockOptions;
         private readonly LocalStorageService _sut;
+        private readonly IStorageProvider _storageProvider;
+        private readonly IJsonSerializer _serializer;
 
-        private static readonly string _key = "testKey";
+        private const string Key = "testKey";
 
         public GetItemAsync()
         {
-            _mockJSRuntime = new Mock<JSRuntimeWrapperAsync>();
-            _mockOptions = new Mock<IOptions<LocalStorageOptions>>();
-            _jsonOptions = new JsonSerializerOptions();
-            _jsonOptions.Converters.Add(new TimespanJsonConverter());
-            _mockOptions.Setup(u => u.Value).Returns(new LocalStorageOptions());
-            _sut = new LocalStorageService(_mockJSRuntime.Object, _mockOptions.Object);
+            var mockOptions = new Mock<IOptions<LocalStorageOptions>>();
+            var jsonOptions = new JsonSerializerOptions();
+            jsonOptions.Converters.Add(new TimespanJsonConverter());
+            mockOptions.Setup(u => u.Value).Returns(new LocalStorageOptions());
+            _serializer = new SystemTextJsonSerializer(mockOptions.Object);
+            _storageProvider = new InMemoryStorageProvider();
+            _sut = new LocalStorageService(_storageProvider, _serializer);
         }
 
         [Theory]
-        [InlineData("stringTest")]
-        [InlineData(11)]
-        [InlineData(11.11)]
-        public async Task Should_DeserialiseToCorrectType<T>(T value)
+        [InlineData("")]
+        [InlineData("  ")]
+        [InlineData(null)]
+        public void ThrowsArgumentNullException_When_KeyIsInvalid(string key)
+        {
+            // arrange / act
+            var action = new Func<Task>(async () => await _sut.GetItemAsync<object>(key));
+
+            // assert
+            Assert.ThrowsAsync<ArgumentNullException>(action);
+        }
+        
+        [Theory]
+        [InlineData("Item1", "stringTest")]
+        [InlineData("Item2", 11)]
+        [InlineData("Item3", 11.11)]
+        public async Task ReturnsDeserializedDataFromStore<T>(string key, T data)
         {
             // Arrange
-            var serialisedData = "";
-            if (typeof(T) == typeof(string))
-                serialisedData = value.ToString();
-            else
-                serialisedData = JsonSerializer.Serialize(value, _jsonOptions);
-
-            _mockJSRuntime.Setup(x => x.InvokeAsync<string>("localStorage.getItem", new[] { _key }))
-                          .Returns(() => new ValueTask<string>(serialisedData));
-
+            await _sut.SetItemAsync(key, data);
+            
             // Act
-            var result = await _sut.GetItemAsync<T>(_key);
+            var result = await _sut.GetItemAsync<T>(key);
 
             // Assert
-            Assert.Equal(value, result);
-            _mockJSRuntime.Verify();
+            Assert.Equal(data, result);
         }
 
         [Fact]
-        public async Task Should_DeserialiseValueToNullableInt()
+        public async Task ReturnsComplexObjectFromStore()
         {
             // Arrange
-            int? value = 6;
-            var serialisedData = JsonSerializer.Serialize(value, _jsonOptions);
-
-            _mockJSRuntime.Setup(x => x.InvokeAsync<string>("localStorage.getItem", new[] { _key }))
-                          .Returns(() => new ValueTask<string>(serialisedData));
+            var objectToSave = new TestObject(2, "Jane Smith");
+            await _sut.SetItemAsync(Key, objectToSave);
 
             // Act
-            var result = await _sut.GetItemAsync<int?>(_key);
+            var result = await _sut.GetItemAsync<TestObject>(Key);
 
             // Assert
-            Assert.Equal(value, result);
+            Assert.Equal(objectToSave.Id, result.Id);
+            Assert.Equal(objectToSave.Name, result.Name);
         }
-
+        
         [Fact]
-        public async Task Should_DeserialiseValueToDecimal()
+        public async Task ReturnsNullFromStore_When_NullValueSaved()
         {
             // Arrange
-            decimal value = 6.00m;
-            var serialisedData = JsonSerializer.Serialize(value, _jsonOptions);
-
-            _mockJSRuntime.Setup(x => x.InvokeAsync<string>("localStorage.getItem", new[] { _key }))
-                          .Returns(() => new ValueTask<string>(serialisedData));
+            var valueToSave = (string)null;
+            await _sut.SetItemAsync(Key, valueToSave);
 
             // Act
-            var result = await _sut.GetItemAsync<decimal>(_key);
+            var result = await _sut.GetItemAsync<string>(Key);
 
             // Assert
-            Assert.Equal(value, result);
+            Assert.Null(result);
         }
-
+        
         [Fact]
-        public async Task Should_DeserialiseValueToComplexType()
+        public async Task ReturnsStringFromStore_When_JsonExceptionIsThrown()
         {
             // Arrange
-            TestObject value = new TestObject { Id = 1, Name = "John Smith" };
-            var serialisedData = JsonSerializer.Serialize(value, _jsonOptions);
-
-            _mockJSRuntime.Setup(x => x.InvokeAsync<string>("localStorage.getItem", new[] { _key }))
-                          .Returns(() => new ValueTask<string>(serialisedData));
+            var valueToSave = "[{ id: 5, name: \"Jane Smith\"}]";
+            await _storageProvider.SetItemAsync(Key, valueToSave);
 
             // Act
-            var result = await _sut.GetItemAsync<TestObject>(_key);
+            var result = await _sut.GetItemAsync<string>(Key);
 
             // Assert
-            result.Should().BeEquivalentTo(value);
+            Assert.Equal(valueToSave, result);
         }
     }
 }
