@@ -10,9 +10,9 @@
     <div class="col-md-4">
         <h5>Add Item to local storage</h5>
         <div class="input-group">
-            <input class="form-control" type="text" placeholder="Enter a value" @bind="Name" />
+            <input id="Name" class="form-control" type="text" placeholder="Enter a value" @bind="Name" />
             <div class="input-group-append">
-                <button class="btn btn-primary" @onclick="SaveName">Save</button>
+                <button id="NameButton" class="btn btn-primary" @onclick="SaveName">Save</button>
             </div>
         </div>
     </div>
