@@ -4,15 +4,6 @@ namespace Blazored.LocalStorage.StorageOptions
 {
     public class LocalStorageOptions
     {
-        public JsonSerializerOptions JsonSerializerOptions { get; } = new JsonSerializerOptions
-        {
-            DictionaryKeyPolicy = JsonNamingPolicy.CamelCase,
-            IgnoreNullValues = true,
-            IgnoreReadOnlyProperties = true,
-            PropertyNameCaseInsensitive = true,
-            PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            ReadCommentHandling = JsonCommentHandling.Skip,
-            WriteIndented = false
-        };
+        public JsonSerializerOptions JsonSerializerOptions { get; } = new JsonSerializerOptions();
     }
 }
